Move Server info service into auto-closable class
